+++ packages/app-scripts-lit-element/package.json
@@ -9,7 +9,9 @@
   "dependencies": {
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-proposal-decorators": "^7.8.3",
-    "@babel/preset-typescript": "7.9.0"
+    "@babel/preset-typescript": "7.9.0",
+    "@open-wc/testing-karma": "^3.3.19",
+    "deepmerge": "^4.2.2"
   },
   "gitHead": "YYY"
 }
