+++ templates/app-template-lit-element-typescript/package.json
@@ -12,10 +12,12 @@
     "prepare": "snowpack",
     "start": "snowpack dev",
     "build": "snowpack build",
-    "test": "echo \"This template does not include a test runner by default.\" && exit 1"
+    "test": "tsc && karma start"
   },
   "devDependencies": {
+    "@open-wc/testing-karma": "3.3.19",
     "@snowpack/app-scripts-lit-element": "^0.3.4",
+    "deepmerge": "4.2.2",
     "lit-html": "^1.2.1",
     "parcel-bundler": "^1.12.4",
     "snowpack": "^2.0.0-0",
